Renamed wallet util file to beacon
