+++ V2/tezex/src/functions/beacon.ts
